+++ Feedbridge/Views/Dashboard/DashboardView.swift
@@ -1,12 +1,20 @@
+//
+//  DashboardView.swift
+//  Feedbridge
+//
+//  Created by Shreya D'Souza on 3/5/25.
+//
+
 import Charts
 import SpeziAccount
 import SwiftUI
 
+/// Dashboard view displaying baby data such as weights, feeds, wet diapers, and stools.
 struct DashboardView: View {
+    
     @Environment(Account.self) private var account: Account?
     @Environment(FeedbridgeStandard.self) private var standard
     @Binding var presentingAccount: Bool
-    
     @AppStorage(UserDefaults.selectedBabyIdKey) private var selectedBabyId: String?
     @State private var isLoading = true
     @State private var errorMessage: String?
@@ -15,6 +23,7 @@ struct DashboardView: View {
     var body: some View {
         NavigationStack {
             Group {
+                // Show loading, error, or main content
                 if isLoading {
                     ProgressView()
                 } else if let error = errorMessage {
@@ -36,6 +45,7 @@ struct DashboardView: View {
         }
     }
     
+    /// Main content of the dashboard, displaying summary views.
     @ViewBuilder private var mainContent: some View {
         ScrollView {
             VStack(spacing: 16) {
@@ -49,6 +59,8 @@ struct DashboardView: View {
             .padding()
         }
     }
+    
+    /// Loads baby data asynchronously.
     private func loadBaby() async {
         guard let babyId = selectedBabyId else {
             baby = nil
@@ -68,7 +80,7 @@ struct DashboardView: View {
     }
 }
 
-// Define the enum for chart types
+/// Enum for chart types.
 enum ChartType: Identifiable {
     case weight
     case dehydration
