+++ Feedbridge/Views/Dashboard/StoolsView.swift
@@ -1,27 +1,29 @@
 //
-//  WetDiapersView.swift
+//  StoolsView.swift
 //  Feedbridge
 //
 //  Created by Shreya D'Souza on 3/5/25.
 //
 import Charts
 import SwiftUI
-struct WetDiapersView: View {
+
+/// View displaying stool entries in a list and chart.
+struct StoolsView: View {
     @Environment(\.presentationMode) var presentationMode
-    let entries: [WetDiaperEntry]
+    let entries: [StoolEntry]
 
     var body: some View {
         NavigationStack {
-            WetDiaperChart(entries: entries, isMini: false)
+            StoolChart(entries: entries, isMini: false)
                 .frame(height: 300)
                 .padding()
-            wetDiaperEntriesList
+            stoolEntriesList
         }
-        .navigationTitle("Voids")
+        .navigationTitle("Stools")
     }
 
-    // List of Wet Diaper Entries
-    private var wetDiaperEntriesList: some View {
+    /// List of stool entries sorted by date, showing volume, color, and time.
+    private var stoolEntriesList: some View {
         List(entries.sorted(by: { $0.dateTime > $1.dateTime })) { entry in
             VStack(alignment: .leading) {
                 Text("\(entry.volume.rawValue.capitalized) and \(entry.color.rawValue.capitalized)")
